requirements.txt
python-telegram-bot==21.0.1
colorama==0.4.6